Fix bundling so that paths to FLTK libraries are always correct.
The previous attempt at bundling had a bug where the bundled FLTK would build in
a temp directory and correctly link the app the first time. But any subsequent
rebuild would fail to find the FLTK libraries because the bundled fltk-config
pointed to the temp directory but which no longer exists since Cabal/Stack
removes it after the first build.

The answer is to build *and* install the bundled FLTK library to it's final
resting place in the app's '.stack-work/install/...' directory before the
configure step and then set PATH to point to *installed* fltk-config so that
when it is used to generate 'fltkhs.buildinfo', 'fltk-config --ldstaticflags'
points to the app's directory instead of the temporary one in which 'fltkhs' is
being built currently.